+++ apps/docs/content/components/popover/title-props.ts
@@ -1,0 +1,29 @@
+const App = `import {Popover, PopoverTrigger, PopoverContent, Button} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Popover placement="right">
+      <PopoverTrigger>
+        <Button>Open Popover</Button>
+      </PopoverTrigger>
+      <PopoverContent>
+        {(titleProps) => (
+          <div className="px-1 py-2">
+            <h3 className="text-sm font-bold" {...titleProps}>
+              Popover Content
+            </h3>
+            <div className="text-xs">This is a content of the popover</div>
+          </div>
+        )}
+      </PopoverContent>
+    </Popover>
+  );
+}`;
+
+const react = {
+  "/App.jsx": App,
+};
+
+export default {
+  ...react,
+};