+++ packages/hooks/use-scroll-position/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nextui-org/use-scroll-position
 
+## 0.0.0-dev-v2-20230605023127
+
+### Patch Changes
+
+- Popover and Tooltip API simplified
+
 ## 0.0.0-dev-v2-20230604204500
 
 ### Patch Changes